+++ CHANGELOG.md
@@ -30,6 +30,12 @@ All versions prior to 0.9.0 are untracked.
    version of `sigstore-python`
    ([#383](https://github.com/sigstore/sigstore-python/pull/383))
 
+* The per-subcommand options `--rekor-url` and `--rekor-root-pubkey` have been
+  moved to the top-level `sigstore` command. Their subcommand forms are unchanged
+  and will continue to work, but will be marked deprecated in a future stable
+  version of `sigstore-python`
+  ([#381](https://github.com/sigstore/sigstore-python/pull/383))
+
 * `sigstore verify github` has been added, allowing for verification of
   GitHub-specific claims within given certificate(s)
   ([#381](https://github.com/sigstore/sigstore-python/pull/381))